refactor mcp generate dto input prep
